+++ src/Validator/DocumentValidator.php
@@ -40,6 +40,7 @@ use GraphQL\Validator\Rules\UniqueEnumValueNames;
 use GraphQL\Validator\Rules\UniqueFragmentNames;
 use GraphQL\Validator\Rules\UniqueInputFieldNames;
 use GraphQL\Validator\Rules\UniqueOperationNames;
+use GraphQL\Validator\Rules\UniqueOperationTypes;
 use GraphQL\Validator\Rules\UniqueVariableNames;
 use GraphQL\Validator\Rules\ValidationRule;
 use GraphQL\Validator\Rules\ValuesOfCorrectType;
@@ -198,6 +199,7 @@ class DocumentValidator
         if (self::$sdlRules === null) {
             self::$sdlRules = [
                 LoneSchemaDefinition::class                  => new LoneSchemaDefinition(),
+                UniqueOperationTypes::class                  => new UniqueOperationTypes(),
                 KnownDirectives::class                       => new KnownDirectives(),
                 KnownArgumentNamesOnDirectives::class        => new KnownArgumentNamesOnDirectives(),
                 UniqueDirectivesPerLocation::class           => new UniqueDirectivesPerLocation(),